+++ main.py
@@ -1,6 +1,6 @@
 # python 3.11.0
 # discord.py 2.4.0
-# pygame 2.5.2
+# Pillow 10.3.0
 import bot
 if __name__ == "__main__":
     bot.runDiscordBot()